refactor(util/lean_path,util/path): separate search path functions
