+++ src/util/init_module.cpp
@@ -8,7 +8,6 @@ Author: Leonardo de Moura
 #include "util/debug.h"
 #include "util/serializer.h"
 #include "util/name.h"
-#include "util/lean_path.h"
 #include "util/thread.h"
 #include "util/memory_pool.h"
 #include "util/fresh_name.h"
@@ -20,12 +19,10 @@ void initialize_util_module() {
     initialize_thread();
     initialize_ascii();
     initialize_name();
-    initialize_lean_path();
     initialize_fresh_name();
 }
 void finalize_util_module() {
     finalize_fresh_name();
-    finalize_lean_path();
     finalize_name();
     finalize_ascii();
     finalize_thread();
