--- src/util/realpath.h
@@ -1,11 +1,0 @@
-/*
-Copyright (c) 2013 Microsoft Corporation. All rights reserved.
-Released under Apache 2.0 license as described in the file LICENSE.
-
-Author: Leonardo de Moura
-*/
-#pragma once
-#include <string>
-namespace lean {
-std::string lrealpath(char const * fname);
-}
